requirements.txt
python-mpv==1.0.1
yt_dlp==2022.1.21
